pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillforge"
version = "0.1.0"
description = "A Python toolkit that enables CrewAI and LangChain agents to be equipped with domain-specific skills using Anthropic's proven skill format"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SkillForge Team"}
]
keywords = ["crewai", "langchain", "agents", "skills", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
]
crewai = [
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
]
langchain = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
]

[project.scripts]
skillforge = "skillforge.cli.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (may require API keys)",
    "slow: marks tests as slow running",
    "validation: marks tests as assumption validation tests",
    "crewai_assumption: marks tests as validating a specific CrewAI assumption",
    "langchain_assumption: marks tests as validating a specific LangChain assumption",
    "general_assumption: marks tests as validating general SkillForge assumptions",
    "benchmark: marks tests as performance benchmark tests",
    "requires_api_key: marks tests as requiring an API key",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
